+++ src/compiler/emitter/mod.rs
@@ -14,5 +14,7 @@ mod lines_3500_4000;
 pub use lines_3500_4000::*;
 mod lines_4000_4500;
 pub use lines_4000_4500::*;
-mod lines_4500_end;
-pub use lines_4500_end::*;
+mod lines_4500_5000;
+pub use lines_4500_5000::*;
+mod lines_5000_end;
+pub use lines_5000_end::*;
